code/requirements.txt
aniso8601==8.0.0
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
marshmallow==3.5.1
PyJWT==1.7.1
pyrsistent==0.16.0
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.9
Werkzeug==0.16.1
zipp==3.1.0
